Add libtcmu2 support
According to Mike's suggestion, rewrite old libtcmu library in
tcmu-runner and name it libtcmu2 with a new model which drops
netlink handle completely. Once a target is created in targetcli,
the TCMU handler or daemon using this new library will directly
find the uio device in /sys directory instead of handling netlink
messages. By providing a chance for daemon to handle device
create/removal after/before operating in configfs with
targetcli/rtslib we can still reuse rtslib code to process device
management.

The daemon using libtcmu2 should implement 3 functions: CheckConfig,
OpenDevice and CloseDevice. The CheckConfig function is still used
to check configstring. The OpenDevice function is used to find uio
device and manage it. The CloseDevice is used to release device
resources before deleting it by rtslib.

This is the targetcli-fb part changes.

Signed-off-by: Yaowei Bai <baiyaowei@cmss.chinamobile.com>